+++ src/config.ts
@@ -5,7 +5,7 @@ import {
   StickerType,
 } from "./themeTools";
 import { constructSyntax } from "./syntax";
-import { constructCSS } from "./css";
+import {Config, constructCSS} from "./css";
 import path from "path";
 import fs from "fs";
 import os from "os";
@@ -27,6 +27,7 @@ const configFile = path.resolve(configDirectory, ".hyper.doki.config.json");
 export interface DokiThemeConfig {
   themeId: string;
   showSticker: boolean;
+  showWallpaper: boolean;
   stickerType: StickerType;
   useFonts: boolean;
 }
@@ -34,6 +35,7 @@ export interface DokiThemeConfig {
 export const DEFAULT_CONFIGURATION: DokiThemeConfig = {
   themeId: "420b0ed5-803c-4127-97e3-dae6aa1a5972",
   showSticker: true,
+  showWallpaper: true,
   stickerType: StickerType.DEFAULT,
   useFonts: false,
 };
@@ -85,10 +87,10 @@ const getExtraSettings = (): { [key: string]: string } => {
     : {};
 };
 
-export const decorateConfig = (config: any) => {
-  const { theme: dokiTheme } = getTheme();
+export const decorateConfig = (config: Config) => {
+  const { theme: dokiTheme, sticker } = getTheme();
   const syntax = constructSyntax(dokiTheme);
-  const css = constructCSS(dokiTheme);
+  const css = constructCSS(dokiTheme, sticker, config);
   return Object.assign({}, config, syntax, {
     ...getExtraSettings(),
     termCSS: config.termCSS || "",