+++ src/settings.ts
@@ -1,18 +1,21 @@
 import DokiThemeDefinitions from "./DokiThemeDefinitions";
-import { extractConfig, saveConfig } from "./config";
-import { dialog } from "electron";
+import {extractConfig, saveConfig} from "./config";
+import {dialog} from "electron";
 import path from "path";
-import { attemptToUpdateSticker } from "./StickerUpdateService";
-import { StickerType, StringDictonary } from "./themeTools";
+import {attemptToUpdateSticker} from "./StickerUpdateService";
+import {StickerType, StringDictonary} from "./themeTools";
 
 export const SET_THEME = "SET_THEME";
 export const SET_STICKER_TYPE = "SET_STICKER_TYPE";
 export const TOGGLE_STICKER = "TOGGLE_STICKER";
+export const TOGGLE_WALLPAPER = "TOGGLE_WALLPAPER";
 export const TOGGLE_FONT = "TOGGLE_FONT";
 export const STICKER_UPDATED = "STICKER_UPDATED";
 
 function attemptToDoStickerStuff(focusedWindow: any) {
-  attemptToUpdateSticker();
+  attemptToUpdateSticker().then(() => {
+    focusedWindow.rpc.emit(STICKER_UPDATED)
+  });
   setTimeout(() => {
     // triggers event loop to continue download?
     focusedWindow.rpc.emit("refresh");
@@ -28,7 +31,7 @@ const stickerTypeMenus = [
     name: "Secondary",
     type: StickerType.SECONDARY,
   },
-].map(({ name, type }) => {
+].map(({name, type}) => {
   return {
     label: name,
     click: async (_: any, focusedWindow: any) => {
@@ -45,20 +48,20 @@ const stickerTypeMenus = [
 const themes = Object.values(DokiThemeDefinitions)
   .sort((def1, def2) => def1.information.name.localeCompare(def2.information.name))
   .map((dokiDefinition) => {
-  return {
-    label: dokiDefinition.information.name,
-    click: async (_: any, focusedWindow: any) => {
-      saveConfig({
-        ...extractConfig(),
-        themeId: dokiDefinition.information.id,
-      });
-      focusedWindow.rpc.emit(SET_THEME, dokiDefinition);
-      attemptToDoStickerStuff(focusedWindow);
-    },
-  };
-});
+    return {
+      label: dokiDefinition.information.name,
+      click: async (_: any, focusedWindow: any) => {
+        saveConfig({
+          ...extractConfig(),
+          themeId: dokiDefinition.information.id,
+        });
+        focusedWindow.rpc.emit(SET_THEME, dokiDefinition);
+        attemptToDoStickerStuff(focusedWindow);
+      },
+    };
+  });
 
-export const VERSION = "v7.0.0";
+export const VERSION = "v7.1.0";
 const icon = path.resolve(__dirname, "..", "assets", "Doki-Theme.png");
 const showAbout = () => {
   const appName = "Doki Theme";
@@ -131,6 +134,18 @@ export default (providedMenu: any): StringDictonary<any> => {
         },
       },
       {
+        label: "Toggle Wallpaper",
+        click: async (_: any, focusedWindow: Window) => {
+          const savedConfig = extractConfig();
+          const showWallpaper = !(savedConfig.showWallpaper || savedConfig.showWallpaper == undefined);
+          focusedWindow.rpc.emit(TOGGLE_WALLPAPER);
+          saveConfig({
+            ...savedConfig,
+            showWallpaper,
+          });
+        },
+      },
+      {
         label: "Toggle Fonts",
         click: async (_: any, focusedWindow: Window) => {
           const savedConfig = extractConfig();
@@ -147,7 +162,7 @@ export default (providedMenu: any): StringDictonary<any> => {
         label: "View ChangeLog",
         click: async () => {
           // eslint-disable-next-line @typescript-eslint/no-var-requires
-          const { shell } = require("electron");
+          const {shell} = require("electron");
           await shell.openExternal(
             "https://github.com/doki-theme/doki-theme-hyper/blob/master/CHANGELOG.md"
           );